producer/requirements.txt
pandas==2.2.3
kafka-python==2.0.2
